+++ packages/react/src/index.ts
@@ -2,6 +2,6 @@ import Active from './Active';
 import Block from './Block';
 import curious from './curious';
 import Link from './Link';
-import Navigator from './Navigator';
+import CuriBase from './CuriBase';
 
-export { Active, Block, curious, Link, Navigator };
+export { Active, Block, curious, Link, CuriBase };