+++ packages/react/types/CuriBase.d.ts
@@ -4,17 +4,17 @@ import PropTypes from 'prop-types';
 import { CuriConfig, Response } from '@curi/core';
 import { CuriContext } from './interface';
 import { Action } from '@hickory/root';
-export interface NavigatorProps {
+export interface CuriBaseProps {
     config: CuriConfig;
     render: (r: Response, action: string, c?: CuriConfig) => React.ReactElement<any>;
     response: Response;
     action: Action;
 }
-declare class Navigator extends React.Component<NavigatorProps> {
+declare class CuriBase extends React.Component<CuriBaseProps> {
     static childContextTypes: {
         curi: PropTypes.Requireable<any>;
     };
     getChildContext(): CuriContext;
     render(): React.ReactElement<any>;
 }
-export default Navigator;
+export default CuriBase;