+++ packages/react/types/index.d.ts
@@ -2,5 +2,5 @@ import Active from './Active';
 import Block from './Block';
 import curious from './curious';
 import Link from './Link';
-import Navigator from './Navigator';
-export { Active, Block, curious, Link, Navigator };
+import CuriBase from './CuriBase';
+export { Active, Block, curious, Link, CuriBase };